Maintenance
Removed files that will never be used in this project
Cleaned up premake file
Fixed redundant cbuffer array in DXRBase
Removed 35 compile warnings
Updated some submodules

+++ Sail/src/API/DX12/renderer/DX12GBufferRenderer.cpp
@@ -55,10 +55,10 @@ void DX12GBufferRenderer::present(PostProcessPipeline* postProcessPipeline, Rend
 	std::future<void> fut[MAX_RECORD_THREADS];
 
 	int mainThreadIndex = nThreadsToUse - 1;
-	int commandsPerThread = round(count / (float)nThreadsToUse);
+	int commandsPerThread = (int)round(count / (float)nThreadsToUse);
 	int start = 0;
 
-	for (size_t i = 0; i < mainThreadIndex; i++) {
+	for (int i = 0; i < mainThreadIndex; i++) {
 		fut[i] = Application::getInstance()->pushJobToThreadPool(
 			[this, postProcessPipeline, i, frameIndex, start, commandsPerThread, count, nThreadsToUse](int id) {
 			return this->recordCommands(postProcessPipeline, i, frameIndex, start, (i < nThreadsToUse - 1) ? commandsPerThread : commandsPerThread + 1, count, nThreadsToUse);
@@ -86,7 +86,7 @@ void DX12GBufferRenderer::present(PostProcessPipeline* postProcessPipeline, Rend
 #endif // MULTI_THREADED_COMMAND_RECORDING
 }
 
-void DX12GBufferRenderer::recordCommands(PostProcessPipeline* postProcessPipeline, const int threadID, const int frameIndex, const int start, const int nCommands, size_t oobMax, int nThreads) {
+void DX12GBufferRenderer::recordCommands(PostProcessPipeline* postProcessPipeline, const int threadID, const int frameIndex, const int start, const int nCommands, unsigned int oobMax, int nThreads) {
 	assert(!postProcessPipeline); // Post process not allowed on the GBuffer pass!
 
 	auto& allocator = m_command[threadID].allocators[frameIndex];
